@@ -4,11 +4,17 @@
   "private": true,
   "description": "EduLaw is a project that assists parents in reporting violations of their children's educational rights to the appropriate legal entity.",
   "scripts": {
-    "dev": "next dev",
+    "dev": "npm run db:start && next dev",
+    "dev2": "npm run && next dev",
+    "dev:frontend": "next dev",
+    "stop": "npm run db:stop",
+    "db:start": "docker-compose -p edlaw -f docker-compose.yml --env-file .env up -d",
+    "db:stop": "docker-compose  -p edlaw -f docker-compose.yml --env-file .env down",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "test:watch": "jest --watch",
     "prepare": "husky install"
   },
   "repository": {
@@ -30,16 +36,27 @@
     "@mui/material": "^5.0.4",
     "@mui/styled-engine-sc": "^5.0.3",
     "@mui/styles": "^5.0.1",
+    "@types/csv-parse": "^1.2.2",
     "@types/material-ui": "^0.21.9",
+    "bcrypt": "^5.0.1",
+    "bcryptjs": "^2.4.3",
+    "csv-parse": "^5.0.4",
     "formik": "^2.2.9",
     "jspdf": "^2.4.0",
-    "next": "11.1.2",
+    "mongodb": "^4.4.1",
+    "mongoose": "^6.2.8",
+    "next": "^12.1.1",
+    "next-auth": "^4.2.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "styled-components": "^5.3.3",
     "yup": "0.32.11"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
+    "@types/bcryptjs": "^2.4.2",
+    "@types/jest": "^27.4.1",
     "@types/react": "17.0.24",
     "@types/styled-components": "^5.1.17",
     "@typescript-eslint/eslint-plugin": "^5.3.1",
@@ -49,6 +66,7 @@
     "eslint-config-next": "11.1.2",
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.4",
+    "jest": "^27.5.1",
     "lint-staged": "^11.2.6",
     "prettier": "2.4.1",
     "typescript": "4.4.3"
