@@ -1,62 +1,97 @@
 import { FieldHookConfig, useField } from 'formik'
 import React, { ChangeEvent } from 'react'
 import { Answer } from '../../models'
-import { FormAnswer } from '../../utils/FormContext'
-import { RadioButton } from '../../components/FormStyles/RadioButton'
-import { QuestionText } from '../FormStyles/QuestionText'
+import { RadioFormAnswer } from '../../utils/FormContext'
+import {
+  RadioButton,
+  RadioInputCheckedIcon,
+  RadioInputIcon,
+} from '../../components/FormStyles/RadioButton'
 import QuestionLayout from '../FormStyles/QuestionLayout'
+import Radio from '@mui/material/Radio'
+import RadioGroup from '@mui/material/RadioGroup'
+import FormControlLabel from '@mui/material/FormControlLabel'
+import styled from 'styled-components'
+
+const StyledRadioText = styled.body`
+  font-size: 14px;
+`
+
+const StyledFormControlLabel = styled(FormControlLabel)`
+  span {
+    :hover {
+      background-color: transparent;
+    }
+  }
+`
 
 interface MyRadioProps {
   name: string
   label: string
   options: Answer[]
   onChange: (event: ChangeEvent<HTMLInputElement>) => void
-  ans?: FormAnswer
+  ans?: RadioFormAnswer
+  tooltip?: { tooltipText: string; tooltipHoveredText: string }
 }
 
-export const MyRadio: React.FC<MyRadioProps & FieldHookConfig<string>> = ({
-  ...props
-}): JSX.Element => {
+export const MyRadio: React.FC<MyRadioProps & FieldHookConfig<string>> = (
+  props
+): JSX.Element => {
   const [field, meta] = useField(props)
+  const { ans, name, label, options, onChange, tooltip } = props
   // renders input type radio, determines whether or not it should be checked initially
-  function initialRadio(optionId: number): JSX.Element {
-    if (props.ans) {
-      if (optionId.toString() === props.ans.answerId) {
-        return (
-          <RadioButton
-            type="radio"
-            {...field}
-            value={optionId}
-            onChange={props.onChange}
-            defaultChecked
-          />
-        )
-      }
-    }
+  function renderButtonRadio(option: Answer, optionId: number): JSX.Element {
     return (
-      <RadioButton
-        type="radio"
-        {...field}
+      <div key={option.content}>
+        <RadioButton
+          type="radio"
+          {...field}
+          value={optionId}
+          onChange={onChange}
+          defaultChecked={ans && optionId === ans.answerId}
+        />
+        <label>{option.content}</label>
+      </div>
+    )
+  }
+  function renderLongRadio(option: Answer, optionId: number) {
+    return (
+      <StyledFormControlLabel
+        style={{ marginLeft: 0 }}
         value={optionId}
-        onChange={props.onChange}
+        key={option.content}
+        label={<StyledRadioText>{option.content || ''}</StyledRadioText>}
+        control={
+          <Radio
+            icon={<RadioInputIcon />}
+            checkedIcon={<RadioInputCheckedIcon />}
+            disableRipple
+            {...field}
+            onChange={onChange}
+            value={optionId}
+          />
+        }
       />
     )
   }
 
-  const answers = props.options.map(function (option, i) {
-    return (
-      <div key={option.content}>
-        {initialRadio(i)}
-        <label>{option.content}</label>
-      </div>
+  const renderRadioAnswers = (options: Answer[]) => {
+    return options.some(({ content }) => content && content.length < 10) ? (
+      <>{options.map(renderButtonRadio)}</>
+    ) : (
+      <RadioGroup style={{ gap: 8 }} defaultValue={ans && ans.answerId}>
+        {/* Reverse is called here so that y */}
+        {options.slice().reverse().map(renderLongRadio)}
+      </RadioGroup>
     )
-  })
+  }
 
   return (
-    <div key={props.name} role="group" aria-labelledby="my-radio-group">
+    <div key={name} role="group" aria-labelledby="my-radio-group">
       <QuestionLayout
-        questionText={<QuestionText>{props.label}</QuestionText>}
-        input={<>{answers}</>}
+        questionText={label}
+        tooltip={tooltip}
+        input={renderRadioAnswers(options)}
       />
 
       {meta.touched && meta.error ? (
