@@ -1,15 +1,18 @@
 import styled from 'styled-components'
+import { QuestionsWithBlockText } from './QuestionText'
+import Tooltip from '../DynamicForm/Tooltip'
 
 interface QuestionProps {
-  questionText: JSX.Element
   input: JSX.Element
+  questionText: string
+  tooltip?: { tooltipText: string; tooltipHoveredText: string }
+  results?: JSX.Element[]
 }
-
 const QuestionDiv = styled.div`
   display: flex;
   flex-direction: column;
   margin-bottom: 40px;
-  gap: 40px;
+  gap: 20px;
 `
 
 const InputDiv = styled.div`
@@ -19,10 +22,17 @@ const InputDiv = styled.div`
 `
 
 // how to attach this to formik? hopefully input will already be attached...
-const QuestionLayout: React.FC<QuestionProps> = ({ questionText, input }) => {
+const QuestionLayout: React.FC<QuestionProps> = ({
+  questionText,
+  tooltip,
+  results,
+  input,
+}) => {
   return (
     <QuestionDiv>
-      {questionText}
+      {results}
+      <QuestionsWithBlockText questionText={questionText} />
+      <Tooltip tooltip={tooltip} />
       <InputDiv>{input}</InputDiv>
     </QuestionDiv>
   )
