@@ -1,27 +1,20 @@
 import React from 'react'
 import Divider from '@mui/material/Divider'
+import { MainDiv } from './LandingStyles'
 
 function LandingContent() {
   return (
     <div className="main">
-      <div
-        style={{
-          width: '90%',
-          marginLeft: '20px',
-          marginTop: '30px',
-          marginBottom: '30px',
-        }}
-      >
+      <MainDiv>
         {/* put this in a stylesheet eventually :D and put the content into a const / cleanup*/}
         <h1>Filing a Complaint with the State Department of Education</h1>
-        <p>Last updated January 2021</p>
         <Divider />
         <br />
         <p>
           The Problem Resolution System (PRS) is the Massachusetts Department of
           Elementary and Secondary Education&apos;s (DESE) system for addressing
           complaints about students&apos; education rights and the legal
-          requirements for education.{' '}
+          requirements for education.
           <em>
             If you think a school or district has violated a student&apos;s
             rights
@@ -36,12 +29,12 @@ function LandingContent() {
         </p>
         <br />
         <p>
-          When you are ready, you can select &quot;Continue&quot; on the right
+          When you are ready, you can select &quot;Continue&quot; at the bottom
           to start your complaint. Please note that this resource only applies
           to students in Massachusetts.
         </p>
         <br /> <br />
-      </div>
+      </MainDiv>
     </div>
   )
 }
