@@ -6,12 +6,14 @@ import ListItemText from '@mui/material/ListItemText'
 import Collapse from '@mui/material/Collapse'
 import Divider from '@mui/material/Divider'
 import ExpandMore from '@mui/icons-material/ExpandMore'
+import ExpandLessIcon from '@mui/icons-material/ExpandLess'
 import SubMenuItem from './SubMenuItem'
 import Link from 'next/link'
+import { COLORS } from '../../constants/colors'
 
 enum MENU_OPTS {
-  STUDENT_RIGHTS = '/studentrights',
-  ABOUT_PRS = '/aboutprs',
+  INFO = '/info',
+  ABOUT_PRS = '/prs',
 }
 
 function RightsPrsMenu() {
@@ -22,11 +24,11 @@ function RightsPrsMenu() {
     return (
       <Link href={link} passHref>
         <ListItemButton sx={{ width: '100%' }}>
+          {pathname === link ? <ExpandLessIcon /> : <ExpandMore />}
           <ListItemText
             primary={label}
             sx={{ fontSize: '22px', fontWeight: 400, pl: 1 }}
           />
-          <ExpandMore />
         </ListItemButton>
       </Link>
     )
@@ -36,31 +38,24 @@ function RightsPrsMenu() {
     <List
       sx={{
         width: '100%',
-        background: '#F4F5F7',
+        background: COLORS.LIGHT_GREY,
         padding: '4px',
         position: 'sticky',
-        top: '4px',
+        top: '15px',
       }}
     >
-      {MainListItem('Student Rights', MENU_OPTS.STUDENT_RIGHTS)}
-      <Collapse
-        in={pathname === MENU_OPTS.STUDENT_RIGHTS}
-        timeout="auto"
-        unmountOnExit
-      >
+      {MainListItem('General Information', MENU_OPTS.INFO)}
+      <Collapse in={pathname === MENU_OPTS.INFO} timeout="auto" unmountOnExit>
         <List component="div" disablePadding>
           <SubMenuItem
             label={'School Discipline & Education'}
-            link={`${MENU_OPTS.STUDENT_RIGHTS}#school-discipline`}
+            link={`${MENU_OPTS.INFO}#school-discipline`}
           />
           <SubMenuItem
             label={'Special Education'}
-            link={`${MENU_OPTS.STUDENT_RIGHTS}#special-education`}
+            link={`${MENU_OPTS.INFO}#special-education`}
           />
-          <SubMenuItem
-            label={'Bullying'}
-            link={`${MENU_OPTS.STUDENT_RIGHTS}#bullying`}
-          />
+          <SubMenuItem label={'Bullying'} link={`${MENU_OPTS.INFO}#bullying`} />
         </List>
       </Collapse>
       <Divider />
