@@ -1,4 +1,5 @@
 import React from 'react'
+import { QuestionType } from '../models/question'
 
 export interface FormValues {
   formAnswers: {
@@ -6,12 +7,32 @@ export interface FormValues {
   }
 }
 
-export interface FormAnswer {
-  questionId: string
-  question?: string
-  answerId: string
+export type FormAnswer =
+  | RadioFormAnswer
+  | ContinueQuestionAnswer
+  | TextFormAnswer
+
+interface QuestionAnswer {
+  questionId: number
+}
+
+export interface ContinueQuestionAnswer extends QuestionAnswer {
+  type: QuestionType.CONTINUE
+}
+
+export interface RadioFormAnswer extends QuestionAnswer {
+  type: QuestionType.RADIO
+  answerId: number
+}
+export interface TextFormAnswer extends QuestionAnswer {
+  type: QuestionType.TEXT
+  userAnswer: string
+}
+
+export interface FormResult {
+  question: string
   answer?: string
-  userAnswer?: string
+  formAnswer: FormAnswer
 }
 
 // Interface has two parts
@@ -20,9 +41,7 @@ export interface FormContextInterface {
   updateFormValues?: (formValues: FormValues) => void
 }
 
-export const emptyFormValues: FormValues = {
-  formAnswers: {},
-}
+export const emptyFormValues: FormValues = { formAnswers: {} }
 
 export const defaultFormState: FormContextInterface = {
   formValues: emptyFormValues,
